refactor(05-02): relocate strategies, framework warmup short-circuit, enum emit (D-13/D-15/D-04)
- move SMA_MACD_strategy + empty_strategy into strategy_handler/strategies/
- adopt config-object constructors; remove in-strategy warmup guard + __str__/__repr__
- handler short-circuits on strategy.warmup before generate_signal (D-15)
- emit order_type as OrderType enum directly; drop OrderType() boundary parse (D-04)
- migrate run_backtest.py construction to SMA_MACDConfig; update all import sites
- oracle byte-exact: 134 trades / final_equity 46189.87730727451 (HARD-04)
