+++ tests/integration/test_reservation_inertness.py
@@ -66,7 +66,7 @@ def traced_run(tmp_path_factory):
 
     module = _load_run_backtest_module()
 
-    from itrader.strategy_handler.SMA_MACD_strategy import SMA_MACD_strategy
+    from itrader.strategy_handler.strategies.SMA_MACD_strategy import SMA_MACD_strategy
     from itrader.trading_system.backtest_trading_system import TradingSystem
 
     system = TradingSystem(
